fix(stsauth): Only parse input fields for the loginForm form to submit
the mfa login page included two username fields (username and UserName) in two different forms. the UserName field was resulting in a 500 error from the upstream server. this change only submits input fields from the loginForm form in the UI. this change also adds error handling for when the login response is >=400 and removes an unused dependency on the datetime module